requirements.txt
djangorestframework
Pillow
psycopg2-binary
python-decouple
celery
redis
drf_yasg
django-filter
django-channels
django-rest-knox
djangorestframework-simplejwt